+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * ECS - Display only assigned for a user affiliations in the lookup. Refs UIF-561.
 * *BREAKING* Create settings page to define alternate exchange rate source. Refs UIF-581.
 * *BREAKING* Update `orders` and `order-lines` interfaces. Refs UIF-581.
+* Prevent removal of locations from tenants not associated with the current user. Refs UIF-606.
 * Add a new field `providerType` to exchange rate source settings. Refs UIF-605.
 
 ## [8.0.2](https://github.com/folio-org/ui-finance/tree/v8.0.2) (2025-04-11)